power: ignore requests for uninitialized devices
Because this driver uses lazy initialization, if the GPU is never given
a task it will not be initialized. Even if the device is not fully
initialized it is still registered with the DRM subsystem and capable of
receiving OS APM signals.

As such these events should not result in a fatal error by default. They
should, however, still be logged in debug builds. Outside of the
behavior described above, it is normally the result of a fault to make
it to this section of code with an uninitialized device.

Signed-off-by: Randolph Sapp <rs@ti.com>