pvr_drm: conditionally generate driver-specific keys
This adds a check for the pvr_connection pointer before populating
driver specific keys.

When showing driver specific fdinfo, the driver does not currently
consider if the fd is actively connected to a pvrsrvkm task before
attempting to populate driver specific keys.

We'll print a message under a debug build, as this behavior can be
unusual under certain circumstances, but we really can't control what
userspace does with these descriptors.

Signed-off-by: Randolph Sapp <rs@ti.com>