buildvars: macro prefix kernel directory
Force a clean path for the KERNELDIR variable and use the standard macro
and debug prefixes for it as well. Note that it does not include a
trailing slash. This is intentional, considering some paths like
dwarf's "DW_AT_comp_dir" will use the current directory without a
trailing slash, and the prefix variables are vary rudimentary.

For whatever reason, dwarfv5 is ignoring the debug-prefix-map argument
with GCC 11.5.0 and leaving a reference to the build directory in the
".debug_line_str" section of produced objects, even if the variable is
not reported directly with dwarfdump. Fall back to dwarfv4 for now.

Fixes: 19732104 ("buildvars: fix prefix-maps to prevent references")
Signed-off-by: Randolph Sapp <rs@ti.com>